Feature/addpost (#71)
* Deleted empty file (file_upload.php)

* WIP: feature/addpost

#Edited:
**admin/addpost.php** <-
Started working on handling errors when uploading new blog posts. This
is a work in progress, but in short:

1. Added variables to print values on input fields. If no value, empty
string is printed.

2. Added array $errors, which is used for printing $obligatoryField
when true.

Once this feature is finished, we should remove the possibility to
upload data to the database if any field is empty.

* feature/addpost

#Edited:
**admin/addpost.php** <- Added comments for better structure, renamed
variable $required_fields to $requiredFields. This page now checks if
all fields are filled. If not, error message is printed.

* Replaces switch with just on success message.

* WIP: Feature/addpost

#Edited:
**admin/confirmation.php** <- Added new file, this replaces the GET
success message. Now, when a successful database update is made, user
is redirected to confirmation.php

**admin/addpost.php** <-
1. Replaced _die(header("Location: ./addpost.php?success”));_ with
_Redirect to confirmation.php_

2. Removed var_dump()

3. Removed upload message.

* Added error message

* Removed comments, added header.php again

* Moved around and added if around ($stmt->prepare($updateQuery))