Added new classes to burger menu for styling categories and added them into _base css
Added new classes in header.php in burgermenu. Classes were added to <a
href> tags named hamburger-menu and hamburger-categories and then added
in to base css with comments TODO: We need to discuss if this solution
is optional or not. categories in the meny have a darker green (#1d7f72)
so seperate them from Start, Kontakt and Logga ut.  Comment out the old
class (.hamburger .has-sub ul li a) for the moment and also added TODO
comment.