Rearranged in scss folder
#Ok guys, lots of new stuff here!
#Edited:
**index.php** <- Changed name on two div classes, nothing major.

**styles/css/normalize.css** <- Moved normalize to scss folder, so that
it is now included in our css.

**styles/scss/_alerts.scss** <- New file, only contains error-msg, but
could contain success-msg et cetera. You get the picture.

**styles/scss/_base.scss** <- Heavily reduced file since I’ve now tried
to arrange our scss files in a component way.

**styles/scss/_buttons.scss** <- This is where we keep all styles
regarding buttons.

**styles/scss/_comments.scss** <-This is where we keep all styles
regarding comments.

**styles/scss/_form.scss** <- Styles regarding forms, like input
fields, goes here.

**styles/scss/_hamburger.scss** <- The famous hamburger!

**styles/scss/_typography.scss** <- If you need to edit typography,
mainly do it in this file.

**styles/scss/_utilities.scss** <- Utilities are those classes that are
not specific to any certain object (like .hidden, or .dark). Check it
out!

**styles/scss/_variables.scss** <- All our variables, neatly arranged
in this single file.

**styles/scss/main.scss** <- Is now updated with all the files to
combine. The order in this file is important, lets talk about it.